Enable emojify for index + id relevant els
Setting `emojify: true` in the preamble loads the CSS and JS assets and
triggers `emojify.run()` which excludes content within anchors. As post
listings are wrapped inside anchors, we effectively re-run `emojify.run`
with the relevant elements (header and paragraph) such that they can be
emojified.

This approach effectively copied the content from
_includes/emojify-script.html and lacks elegance but is pragmatic enough
for now.